Restructure package to be go get'able with standard Go ./cmd/<binary>/... commands and internal/ package
