+++ internal/version.go
@@ -1,0 +1,3 @@
+package internal
+
+var Version string
